refactor(oracle): migrate icmp and icmpv6 to per-protocol plugins
